+++ lib/wanderer_app/character/tracker_pool.ex
@@ -25,9 +25,6 @@ defmodule WandererApp.Character.TrackerPool do
   @update_ship_interval :timer.seconds(2)
   @update_info_interval :timer.minutes(1)
   @update_wallet_interval :timer.minutes(1)
-  @inactive_character_timeout :timer.minutes(5)
-
-  @pause_tracking_timeout :timer.minutes(60 * 24)
 
   @logger Application.compile_env(:wanderer_app, :logger)
 
@@ -55,12 +52,6 @@ defmodule WandererApp.Character.TrackerPool do
 
     tracked_ids
     |> Enum.each(fn id ->
-      # WandererApp.Cache.put(
-      #   "character:#{id}:tracking_paused",
-      #   true,
-      #   ttl: @pause_tracking_timeout
-      # )
-
       Cachex.put(@cache, id, uuid)
     end)
 
@@ -87,12 +78,6 @@ defmodule WandererApp.Character.TrackerPool do
     Registry.update_value(@unique_registry, Module.concat(__MODULE__, uuid), fn r_tracked_ids ->
       [tracked_id | r_tracked_ids]
     end)
-
-    # WandererApp.Cache.put(
-    #   "character:#{tracked_id}:tracking_paused",
-    #   true,
-    #   ttl: @pause_tracking_timeout
-    # )
 
     # Cachex.get_and_update(@cache, :tracked_characters, fn ids ->
     #   {:commit, ids ++ [tracked_id]}
@@ -236,13 +221,17 @@ defmodule WandererApp.Character.TrackerPool do
       characters
       |> Task.async_stream(
         fn character_id ->
-          WandererApp.TaskWrapper.start_link(
-            WandererApp.Character.Tracker,
-            :check_offline,
-            [
-              character_id
-            ]
-          )
+          if WandererApp.Character.can_pause_tracking?(character_id) do
+            WandererApp.TaskWrapper.start_link(
+              WandererApp.Character.Tracker,
+              :check_offline,
+              [
+                character_id
+              ]
+            )
+          else
+            :ok
+          end
         end,
         timeout: :timer.seconds(15),
         max_concurrency: System.schedulers_online(),