+++ lib/wanderer_app/env.ex
@@ -16,6 +16,18 @@ defmodule WandererApp.Env do
   def invites, do: get_key(:invites, false)
   def map_subscriptions_enabled?, do: get_key(:map_subscriptions_enabled, false)
   def public_api_disabled?, do: get_key(:public_api_disabled, false)
+
+  @decorate cacheable(
+              cache: WandererApp.Cache,
+              key: "active_tracking_pool"
+            )
+  def active_tracking_pool, do: get_key(:active_tracking_pool, "default")
+
+  @decorate cacheable(
+              cache: WandererApp.Cache,
+              key: "tracking_pool_max_size"
+            )
+  def tracking_pool_max_size, do: get_key(:tracking_pool_max_size, 300)
   def character_tracking_pause_disabled?, do: get_key(:character_tracking_pause_disabled, true)
   def character_api_disabled?, do: get_key(:character_api_disabled, false)
   def wanderer_kills_service_enabled?, do: get_key(:wanderer_kills_service_enabled, false)
@@ -24,7 +36,7 @@ defmodule WandererApp.Env do
   def admin_username, do: get_key(:admin_username)
   def admin_password, do: get_key(:admin_password)
   def corp_wallet, do: get_key(:corp_wallet, "")
-  def corp_wallet_eve_id, do: get_key(:corp_wallet_eve_id, -1)
+  def corp_wallet_eve_id, do: get_key(:corp_wallet_eve_id, "-1")
   def corp_eve_id, do: get_key(:corp_id, -1)
   def subscription_settings, do: get_key(:subscription_settings)